Add cosign sign and attest task
Adds an additional "task" to cosign sign the image that was just
built, generate a provenance attestation predicate, and then use it
to cosign attest the image.

Accesses the signing secret and password from environment variables
that need to be created from Jenkins credentials.

Outputs the public key to stdout and also to a task result to make
it accessible.

Also adds a script that can be used to verify an image's signature
and show the attestation to confirm things are working as expected.

The content of the attestation is mainly placeholder right now. Will
refine and improve it later.

Note that the value for finishedOn is not accurate, but I'm not sure
how we can do any better considering we're pushing this while the
pipeline is still running.

Ref: https://issues.redhat.com/browse/EC-743
Ref: https://issues.redhat.com/browse/EC-744