Add script to generate a signing secret for cosign
Use environment variables to store the secrets that are generated
since I think it will be more convenient to manage than cosign.key
and cosign.pub files.

As mentioned in the comments these need to be created as secret text
credentials in Jenkins. This could be perhaps be automated in future
using the Jenkins CLI.

Ref: https://issues.redhat.com/browse/EC-743