Read signing secret env vars when required only
Rather than have them visible for the entire pipeline it's better to
reduce the scope and have them visible only when they're needed for
cosign to do its signing and attesting.

Do a similar change for the verification, but note that not so much
of a problem if public is key become visible since it's not really a
secret that needs to be protected.

Ref: https://issues.redhat.com/browse/EC-740