Use template for the github pipeline also
I didn't test the pipeline yet, but it's quite similar to the
existing file so I think it should work.

The one change to the pipeline is to separate the the dependency
fetching from the init task, see the new step call "Pre-init". I
think that's an improvement, but will need to confirm it works.