Create template for GitLab pipeline
This changes the structure to be more similar to the other
pipelines.

Note there was some extra tweaking needed to ensure the
cosign-sign-attest task doesn't try to run in parallel with the
buildah task. (I considered introducing schema in
templates/data.yaml for doing that kind of thing but decided not.)

Note that because we're using the new base image there's no more
need for before_script.