Do cosign login explicitly in buildah task
The location of the auth.json file is different in my GitLab base
image build on alpine. Looks like it's in /run/containers/0/auth.json

On top of that, the $XDG_RUNTIME_DIR is not set.

Rather than figure out a good way to handle all that, let's just do
an explicit login. Since we have the credentials available, this is
quite easy to do.