cosign-sign-attest: add SBOM_BLOB byproduct
This is required so that the promotion pipeline can download an SBOM
while guaranteeing that it has not been tampered with. The attestation
is signed => we can trust that the SBOM_BLOB uri was generated by a
trusted pipeline. The SBOM_BLOB uri contains the sha256 digest => we
can verify that the content has not been tampered with.

refs:

- https://slsa.dev/spec/v1.0/provenance#schema
- https://github.com/in-toto/attestation/blob/main/spec/v1/resource_descriptor.md

Signed-off-by: Adam Cmiel <acmiel@redhat.com>