download-sbom: support downloading v1.0 Provenance
Previously, we used --type=slsaprovenance, which is an alias for
https://slsa.dev/provenance/v0.2. Trying to download a v1 provenance
would fail with

    error during command execution: none of the attestations matched the
    predicate type: slsaprovenance, found: https://slsa.dev/provenance/v1

Implement slightly hacky support for downloading multiple attestation
types at the same time and use it to try both the v0.2 and v1.0 versions
of the SLSA provenance.

Note: the jq code for extracting SBOM_BLOB_URL already supports multiple
attestations in the same file. (It still needs an update to support the
SBOM_BLOB byproduct in addition to the SBOM_BLOB_URL Tekton result, that
will come later.)

Signed-off-by: Adam Cmiel <acmiel@redhat.com>