gather-deploy-images: specify envs to gather
This essentially ports
https://github.com/konflux-ci/build-definitions/commit/adcd25b64294447a3a742e65679fe12002fa5850

The script now takes arguments that specify which environments it should
gather images from. Passing no arguments is equivalent to passing
`development stage prod` (as before).

This is needed because the promotion pipeline should EC-verify images in
all the environments but should skip the dev env when uploading SBOMs.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>