Set CI_TYPE in the two task runner containers
At this stage it has no effect in the GitHub container, since we
aren't running anything in that container, but I'll set it there
anyway to be consistent.

Note that the var is likely to be overridden/masked by the var set
by the CI config, but I figure there's no harm in setting it in the
Dockerfile anyhow, and I suspect it will make some testing/hacking
workflows easier.