Easier custom quay org and repo for runner image
If you're hacking on the task runner image, you should not push to
the redhat-appstudio repo. This makes it easier to specify your
personal quay org and repo.

There is warning produced when building the base image because the
build args are not consumed, but I think we can live with that.