Run steps sequentially in GitLab by default
In some cases we need to ensure tasks run sequentially, but setting
the "concurrent" field lets them run in parallel like they do now.

Note that currently GitLab is the only CI type that runs the steps
in parallel when "concurrent" is set. Making Jenkins and GitHub
behave that way too is something for the future.