Use runner image for GitHub
The main issue with running a container on GitHub Actions seems to be
with the default AppArmor profile configured -- it does not allow user
namespaces needed by buildah.

Another characteristic of GitHub Actions is that the container
configured for a Workflow job is used for all steps, so we can't run a
command on the host to enable user namespaces. So running something
like:

    sudo sysctl -w kernel.apparmor_restrict_unprivileged_userns=0

is not possible.

This resorts to running the container with extended privileges in order
to allow user namespaces.

Resolves: https://issues.redhat.com/browse/RHTAP-3066