Fix unshare error using runner image in GitHub
The error:
  Error during unshare(CLONE_NEWUSER): Operation not permitted
  time="2024-10-17T21:17:38Z" level=error msg="parsing PID \"\":
  strconv.Atoi: parsing \"\": invalid syntax"
  time="2024-10-17T21:17:[38](https://github.com/simonbaird/rhtap-ci-tester/actions/runs/11393186912/job/31700934153#step:7:39)Z"
  level=error msg="(Unable to determine exit status)"
  Failed buildah login *** for user ***

The main issue with running a container on GitHub Actions seems to be
with the default AppArmor profile configured -- it does not allow user
namespaces needed by buildah.

Another characteristic of GitHub Actions is that the container
configured for a Workflow job is used for all steps, so we can't run a
command on the host to enable user namespaces. So running something
like:

    sudo sysctl -w kernel.apparmor_restrict_unprivileged_userns=0

is not possible.

This resorts to running the container with extended privileges in order
to allow user namespaces.

Resolves: https://issues.redhat.com/browse/RHTAP-3066