Fix git dubious owner errors with GitHub runner
The value of $(pwd) here seems to be /__w/reponame/reponame, which
is different to ${{ github.workspace }}. I suspect it's related to
the volumes mounted in the container, but I'm not sure.

Note there's a similar command in the standard checkout action, but
with the $HOME var temporarily reassigned, so the change is not
permanent.

Ref: https://issues.redhat.com/browse/RHTAP-3066