Refined Jenkins build attestation
Makes the SLSA Provenance attestation produced by RHTAP when running on
Jenkins similar to the attestations produced when running on GitLab and
GitHub by reshuffling and adding several new properties.

Some of the properties are fetched from Jenkins object graph that is not
accessible to the untrusted Pipeline libraries and if run as untrusted
will be blocked and require script approval.

As a result of hat the Pipeline library consisting of `rhtap.groovy` in
`vars/` directory needs to be installed as a global trusted library[1].

Reference: https://issues.redhat.com/browse/EC-774

[1] https://www.jenkins.io/doc/book/pipeline/shared-libraries/#global-shared-libraries